# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * S.S. Jewellers Invoicing API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Profile {
  id: number;
  userId: string;
  shopName: string;
  shopAddress: string;
  gstNumber: string;
  upiId: string;
  state: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  createdAt: string;
}

export interface ProfileInput {
  /** @minLength 1 */
  shopName: string;
  /** @minLength 1 */
  shopAddress: string;
  /** @minLength 1 */
  gstNumber: string;
  /** @minLength 1 */
  upiId: string;
  /** @minLength 1 */
  state: string;
  phone?: string;
  email?: string;
}

export interface InvoiceItem {
  id: number;
  invoiceId: number;
  /** @nullable */
  itemType?: string | null;
  description: string;
  /** @nullable */
  grossWeight?: number | null;
  /** @nullable */
  netWeight?: number | null;
  /** @nullable */
  sellingPricePerGram?: number | null;
  /** @nullable */
  gemstonePrice?: number | null;
  amount: number;
  makingChargePercent: number;
  makingChargeAmount: number;
  gstJewel: number;
  gstMaking: number;
  itemTotal: number;
  /** @nullable */
  hsnCode?: string | null;
}

export interface InvoiceItemInput {
  itemType?: string;
  /** @minLength 1 */
  description: string;
  grossWeight?: number;
  netWeight?: number;
  sellingPricePerGram?: number;
  gemstonePrice?: number;
  amount: number;
  makingChargePercent: number;
  makingChargeAmount: number;
  gstJewel: number;
  gstMaking: number;
  itemTotal: number;
  hsnCode?: string;
}

export interface InvoiceSummary {
  id: number;
  invoiceNumber: string;
  customerName: string;
  /** @nullable */
  customerAddress?: string | null;
  /** @nullable */
  customerGstin?: string | null;
  invoiceDate: string;
  totalAmount: number;
  createdAt: string;
}

export interface Invoice {
  id: number;
  invoiceNumber: string;
  customerName: string;
  /** @nullable */
  customerAddress?: string | null;
  /** @nullable */
  customerGstin?: string | null;
  invoiceDate: string;
  subtotalAmount?: number;
  makingChargesTotal?: number;
  gstJewelTotal?: number;
  gstMakingTotal?: number;
  totalAmount: number;
  createdAt: string;
  items: InvoiceItem[];
}

export interface InvoiceInput {
  /** @minLength 1 */
  customerName: string;
  customerAddress?: string;
  customerGstin?: string;
  invoiceDate: string;
  items: InvoiceItemInput[];
}

export interface DashboardStats {
  totalInvoices: number;
  totalSales: number;
  totalGstCollected: number;
  totalMakingCharges: number;
  thisMonthInvoices: number;
  thisMonthSales: number;
}

export interface MonthlyGst {
  month: number;
  year: number;
  monthName?: string;
  gstJewel: number;
  gstMaking: number;
  totalGst: number;
  invoiceCount: number;
}