# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * S.S. Jewellers Invoicing API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalInvoices: number;
  totalSales: number;
  totalGstCollected: number;
  totalMakingCharges: number;
  thisMonthInvoices: number;
  thisMonthSales: number;
}