# lib/api-zod/src/generated/types/invoice.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * S.S. Jewellers Invoicing API
 * OpenAPI spec version: 0.1.0
 */
import type { InvoiceItem } from './invoiceItem';

export interface Invoice {
  id: number;
  invoiceNumber: string;
  customerName: string;
  /** @nullable */
  customerAddress?: string | null;
  /** @nullable */
  customerGstin?: string | null;
  invoiceDate: string;
  subtotalAmount?: number;
  makingChargesTotal?: number;
  gstJewelTotal?: number;
  gstMakingTotal?: number;
  totalAmount: number;
  createdAt: Date;
  items: InvoiceItem[];
}