# lib/api-zod/src/generated/types/invoiceInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * S.S. Jewellers Invoicing API
 * OpenAPI spec version: 0.1.0
 */
import type { InvoiceItemInput } from './invoiceItemInput';

export interface InvoiceInput {
  /** @minLength 1 */
  customerName: string;
  customerAddress?: string;
  customerGstin?: string;
  invoiceDate: string;
  items: InvoiceItemInput[];
}