# lib/api-zod/src/generated/types/invoiceItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * S.S. Jewellers Invoicing API
 * OpenAPI spec version: 0.1.0
 */

export interface InvoiceItem {
  id: number;
  invoiceId: number;
  /** @nullable */
  itemType?: string | null;
  description: string;
  /** @nullable */
  grossWeight?: number | null;
  /** @nullable */
  netWeight?: number | null;
  /** @nullable */
  sellingPricePerGram?: number | null;
  /** @nullable */
  gemstonePrice?: number | null;
  amount: number;
  makingChargePercent: number;
  makingChargeAmount: number;
  gstJewel: number;
  gstMaking: number;
  itemTotal: number;
  /** @nullable */
  hsnCode?: string | null;
}