# lib/api-zod/src/generated/types/invoiceItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * S.S. Jewellers Invoicing API
 * OpenAPI spec version: 0.1.0
 */

export interface InvoiceItemInput {
  itemType?: string;
  /** @minLength 1 */
  description: string;
  grossWeight?: number;
  netWeight?: number;
  sellingPricePerGram?: number;
  gemstonePrice?: number;
  amount: number;
  makingChargePercent: number;
  makingChargeAmount: number;
  gstJewel: number;
  gstMaking: number;
  itemTotal: number;
  hsnCode?: string;
}