# lib/api-zod/src/generated/types/invoiceSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * S.S. Jewellers Invoicing API
 * OpenAPI spec version: 0.1.0
 */

export interface InvoiceSummary {
  id: number;
  invoiceNumber: string;
  customerName: string;
  /** @nullable */
  customerAddress?: string | null;
  /** @nullable */
  customerGstin?: string | null;
  invoiceDate: string;
  totalAmount: number;
  createdAt: Date;
}