# lib/api-zod/src/generated/types/monthlyGst.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * S.S. Jewellers Invoicing API
 * OpenAPI spec version: 0.1.0
 */

export interface MonthlyGst {
  month: number;
  year: number;
  monthName?: string;
  gstJewel: number;
  gstMaking: number;
  totalGst: number;
  invoiceCount: number;
}