# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * S.S. Jewellers Invoicing API
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  userId: string;
  shopName: string;
  shopAddress: string;
  gstNumber: string;
  upiId: string;
  state: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  createdAt: Date;
}