# lib/api-zod/src/generated/types/profileInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * S.S. Jewellers Invoicing API
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileInput {
  /** @minLength 1 */
  shopName: string;
  /** @minLength 1 */
  shopAddress: string;
  /** @minLength 1 */
  gstNumber: string;
  /** @minLength 1 */
  upiId: string;
  /** @minLength 1 */
  state: string;
  phone?: string;
  email?: string;
}